Move reusable Hawk classes to core app
