+++ datahub/activity_stream/test/test_activity_stream_views.py
@@ -1,22 +1,11 @@
-import datetime
-
 import mohawk
 import pytest
-from freezegun import freeze_time
 from rest_framework import status
 from rest_framework.reverse import reverse
 
 
 def _url():
     return 'http://testserver' + reverse('api-v3:activity-stream:index')
-
-
-def _url_incorrect_domain():
-    return 'http://incorrect' + reverse('api-v3:activity-stream:index')
-
-
-def _url_incorrect_path():
-    return 'http://testserver' + reverse('api-v3:activity-stream:index') + 'incorrect/'
 
 
 def _auth_sender(
@@ -45,59 +34,6 @@ def _auth_sender(
     'get_kwargs,expected_json',
     (
         (
-            # If no X-Forwarded-For header
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender().request_header,
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If second-to-last X-Forwarded-For header isn't whitelisted
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender().request_header,
-                'HTTP_X_FORWARDED_FOR': '9.9.9.9, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the only IP address in X-Forwarded-For is whitelisted
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender().request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the only IP address in X-Forwarded-For isn't whitelisted
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender().request_header,
-                'HTTP_X_FORWARDED_FOR': '123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If third-to-last IP in X-Forwarded-For header is whitelisted
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender().request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4, 124.124.124, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If last of 3 IPs in X-Forwarded-For header is whitelisted
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender().request_header,
-                'HTTP_X_FORWARDED_FOR': '124.124.124, 123.123.123.123, 1.2.3.4',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
             # If the Authorization header isn't passed
             {
                 'content_type': '',
@@ -106,65 +42,13 @@ def _auth_sender(
             {'detail': 'Authentication credentials were not provided.'},
         ),
         (
-            # If the Authorization header generated from an incorrect ID
+            # If the wrong credentials are used
             {
                 'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender(key_id='incorrect').request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the Authorization header generated from an incorrect secret
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender(secret_key='incorrect').request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the Authorization header generated from an incorrect domain
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender(url=_url_incorrect_domain).request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the Authorization header generated from an incorrect path
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender(url=_url_incorrect_path).request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the Authorization header generated from an incorrect method
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender(method='POST').request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the Authorization header generated from an incorrect
-            # content-type
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender(content_type='incorrect').request_header,
-                'HTTP_X_FORWARDED_FOR': '1.2.3.4, 123.123.123.123',
-            },
-            {'detail': 'Incorrect authentication credentials.'},
-        ),
-        (
-            # If the Authorization header generated from incorrect content
-            {
-                'content_type': '',
-                'HTTP_AUTHORIZATION': _auth_sender(content='incorrect').request_header,
+                'HTTP_AUTHORIZATION': _auth_sender(
+                    key_id='incorrect',
+                    secret_key='incorrect',
+                ).request_header,
                 'HTTP_X_FORWARDED_FOR': '1.2.3.4, 123.123.123.123',
             },
             {'detail': 'Incorrect authentication credentials.'},
@@ -173,9 +57,7 @@ def _auth_sender(
 )
 @pytest.mark.django_db
 def test_401_returned(api_client, get_kwargs, expected_json):
-    """If the request isn't properly Hawk-authenticated, then a 401 is
-    returned
-    """
+    """If the request isn't Hawk-authenticated, then a 401 is returned."""
     response = api_client.get(
         _url(),
         **get_kwargs,
@@ -183,51 +65,6 @@ def test_401_returned(api_client, get_kwargs, expected_json):
 
     assert response.status_code == status.HTTP_401_UNAUTHORIZED
     assert response.json() == expected_json
-
-
-@pytest.mark.django_db
-def test_if_61_seconds_in_past_401_returned(api_client):
-    """If the Authorization header is generated 61 seconds in the past, then a
-    401 is returned
-    """
-    past = datetime.datetime.now() - datetime.timedelta(seconds=61)
-    with freeze_time(past):
-        auth = _auth_sender().request_header
-    response = api_client.get(
-        reverse('api-v3:activity-stream:index'),
-        content_type='',
-        HTTP_AUTHORIZATION=auth,
-        HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
-    )
-
-    assert response.status_code == status.HTTP_401_UNAUTHORIZED
-    error = {'detail': 'Incorrect authentication credentials.'}
-    assert response.json() == error
-
-
-@pytest.mark.usefixtures('local_memory_cache')
-@pytest.mark.django_db
-def test_if_authentication_reused_401_returned(api_client):
-    """If the Authorization header is reused, then a 401 is returned"""
-    auth = _auth_sender().request_header
-
-    response_1 = api_client.get(
-        _url(),
-        content_type='',
-        HTTP_AUTHORIZATION=auth,
-        HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
-    )
-    assert response_1.status_code == status.HTTP_200_OK
-
-    response_2 = api_client.get(
-        _url(),
-        content_type='',
-        HTTP_AUTHORIZATION=auth,
-        HTTP_X_FORWARDED_FOR='1.2.3.4, 123.123.123.123',
-    )
-    assert response_2.status_code == status.HTTP_401_UNAUTHORIZED
-    error = {'detail': 'Incorrect authentication credentials.'}
-    assert response_2.json() == error
 
 
 @pytest.mark.django_db
@@ -254,26 +91,7 @@ def test_403_returned(api_client):
 
 
 @pytest.mark.django_db
-def test_empty_object_returned_with_authentication_3_ips(api_client):
-    """If the Authorization and X-Forwarded-For headers are correct,
-    with an extra IP address prepended to the X-Forwarded-For then
-    the correct, and authentic, data is returned
-    """
-    sender = _auth_sender()
-    response = api_client.get(
-        _url(),
-        content_type='',
-        HTTP_AUTHORIZATION=sender.request_header,
-        HTTP_X_FORWARDED_FOR='3.3.3.3, 1.2.3.4, 123.123.123.123',
-    )
-
-    assert response.status_code == status.HTTP_200_OK
-    content = {'secret': 'content-for-pen-test'}
-    assert response.json() == content
-
-
-@pytest.mark.django_db
-def test_empty_object_returned_with_authentication(api_client):
+def test_succeeds_with_valid_redentials(api_client):
     """If the Authorization and X-Forwarded-For headers are correct, then
     the correct, and authentic, data is returned
     """
