+++ datahub/core/test/support/views.py
@@ -1,5 +1,14 @@
+from django.utils.decorators import decorator_from_middleware
 from oauth2_provider.contrib.rest_framework.permissions import IsAuthenticatedOrTokenHasScope
+from rest_framework.response import Response
+from rest_framework.views import APIView
 
+from config.settings.types import HawkScope
+from datahub.core.hawk_receiver import (
+    HawkAuthentication,
+    HawkScopePermission,
+    HawkResponseMiddleware,
+)
 from datahub.core.test.support.models import MultiAddressModel, MyDisableableModel, PermissionModel
 from datahub.core.test.support.serializers import (
     MultiAddressModelSerializer,
@@ -33,3 +42,16 @@ class MultiAddressModelViewset(CoreViewSet):
     permission_classes = []
     serializer_class = MultiAddressModelSerializer
     queryset = MultiAddressModel.objects.all()
+
+
+class HawkView(APIView):
+    """View using Hawk authentication."""
+
+    authentication_classes = (HawkAuthentication,)
+    permission_classes = (HawkScopePermission,)
+    required_hawk_scope = next(iter(HawkScope.__members__.values()))
+
+    @decorator_from_middleware(HawkResponseMiddleware)
+    def get(self, request):
+        """Simple test view with fixed response"""
+        return Response({'content': 'hawk-test-view'})
